+++ resources/views/jenismasalah/index.blade.php
@@ -19,6 +19,21 @@ scratch. This page gets rid of all links and provides the needed markup only.
     @include('Template.sidebar')
   <!-- Content Wrapper. Contains page content -->
   <div class="content-wrapper">
+    @if(session('Data dihapus'))
+    <div class="alert alert-danger" role="alert">
+        {{session('Data dihapus')}}
+    </div>
+    @endif
+    @if(session('Data ditambah'))
+    <div class="alert alert-info" role="alert">
+        {{session('Data ditambah')}}
+    </div>
+    @endif
+    @if(session('Data diedit'))
+    <div class="alert alert-warning" role="alert">
+        {{session('Data diedit')}}
+    </div>
+    @endif
     <!-- Content Header (Page header) -->
     <div class="content-header">
       <div class="container-fluid">
@@ -53,17 +68,17 @@ scratch. This page gets rid of all links and provides the needed markup only.
                     {{-- Data Jenis Masalah --}}
                 </div>
                 <div class="card-body">
-                    <a href="/tambah-jenismasalah" class="btn btn-primary float-right"><i class=" fa fa-plus"></i>Tambah Jenis Masalah</a>
+                    <a href="/tambah-jenismasalah" class="btn btn-success float-right"><i class=" fa fa-plus"></i>Tambah Jenis Masalah</a>
                     <br/>
                     <br/>
-                    <table class="table table-bordered table-hover table-striped">
+                    <table class="table table-bordered table-hover table-striped" id="table-data">
                         <thead>
                             <tr>
                               
                                 {{-- <th>Id_masalah</th> --}}
                                 <th>No</th>
                                 <th>Nama Masalah </th>
-                                <th colspan="3" style="text-align: center;">Aksi</th>
+                                <th style="text-align: center;">Aksi</th>
                             </tr>
                         </thead>
                         <tbody>
@@ -74,7 +89,7 @@ scratch. This page gets rid of all links and provides the needed markup only.
                                 <td>{{ $p->nm_masalah }}</td>
                                 <td>
                                     <a href="/edit-jenismasalah/{{ $p->id_masalah }}" class="btn btn-warning"><i class="fa fa-edit"></i></a>
-                                    <a href="/delete-jenismasalah/{{ $p->id_masalah }}" class="btn btn-danger"><i class="fa fa-trash"></i></a>
+                                    <a href="/delete-jenismasalah/{{ $p->id_masalah }}" class="btn btn-danger"  onclick="return confirm('Yakin anda ingin menghapus data ini?')"><i class="fa fa-trash"></i></a>
                                 </td>
                             </tr>
                             @endforeach
