+++ resources/views/kelas/index.blade.php
@@ -19,6 +19,21 @@ scratch. This page gets rid of all links and provides the needed markup only.
     @include('Template.sidebar')
   <!-- Content Wrapper. Contains page content -->
   <div class="content-wrapper">
+    @if(session('Data dihapus'))
+    <div class="alert alert-danger" role="alert">
+        {{session('Data dihapus')}}
+    </div>
+    @endif
+    @if(session('Data ditambah'))
+    <div class="alert alert-info" role="alert">
+        {{session('Data ditambah')}}
+    </div>
+    @endif
+    @if(session('Data diedit'))
+    <div class="alert alert-warning" role="alert">
+        {{session('Data diedit')}}
+    </div>
+    @endif
     <!-- Content Header (Page header) -->
     <div class="content-header">
       <div class="container-fluid">
@@ -53,10 +68,10 @@ scratch. This page gets rid of all links and provides the needed markup only.
                     {{-- Data Kelas --}}
                 </div>
                 <div class="card-body">
-                    <a href="/tambah-kelas" class="btn btn-primary float-right"><i class=" fa fa-plus"></i>Tambah Kelas</a>
+                    <a href="/tambah-kelas" class="btn btn-success float-right"><i class=" fa fa-plus"></i>Tambah Kelas</a>
                     <br/>
                     <br/>
-                    <table class="table table-bordered table-hover table-striped">
+                    <table class="table table-bordered table-hover table-striped" id="table-data">
                         <thead>
                             <tr>
                                 <th>No</th>
@@ -74,7 +89,7 @@ scratch. This page gets rid of all links and provides the needed markup only.
                                 <td>
                                     <a href="/edit-kelas/{{ $p->kode_kelas }}" class="btn btn-warning"><i class="fa fa-edit"></i></a>
                                     {{-- <a href="/detail-kelas/{{ $p->kode_kelas }}" class="btn btn-success">detail</a> --}}
-                                    <a href="/delete-kelas/{{ $p->kode_kelas }}" class="btn btn-danger"><i class="fa fa-trash"></i></a>
+                                    <a href="/delete-kelas/{{ $p->kode_kelas }}" class="btn btn-danger" onclick="return confirm('Yakin anda ingin menghapus data ini?')"><i class="fa fa-trash"></i></a>
                                 </td>
                             </tr>
                             @endforeach
